Video loader
Uses ffmpeg's http capabilities to remotely probe the video
and extract only the frame we need for thumbnailing. This vastly
cuts down bandwidth between thumbor and its source.

A trade-off due to how thumbor is set up is that we can't
cache the "original" anymore, because we would have do so
on a per-frame basis and purging all requested frames at
once wouldn't be possible without some heavy-duty hacking
on the memcached storage class. This should be acceptable
since videos don't need prerendering that badly, and said
prerendering would be subject to race conditions anyway,
causing parallel requests to the same frame regardless.

Bug: T120205
Change-Id: I3f622181ecb9ad7dd57348a23deb6506a5242798